refactor: SusuBadge로 통합
